// src/app/app-shared/collection/serviceurl.enum.ts
export enum ServicesIDs {
  GetRoot = 0,
  GetLemma = 1,
  SearchInLemma = 2,
  GetLexicalsheetDetailByRoot = 3,
  GetLexicalsheetDetailByLemma = 4,
  GetLexicalsheetDetailById = 5,
  GetCarvingLexicalSheetByRootId = 6,
  GetSummaryEtymologicalLexicalSheetByRootId = 7,
  GetSummaryDocumentDates = 8,
  GetAllCarvingSourceABV = 9,
  GetAllCarvingSource = 10,
  SearchInCorpus = 11,
  SearchByLemmaAutoC = 12,
  GetAllDocumetnts = 13,
  CarvingGetAllDocumetnts = 14,
  GetLemmasCountPerYears = 15,
  GetLemmaCountPerYears = 16,
  GetNextSequence = 17,
  GetPrevSequence = 18,
  GetCountsOfUsageLemmaPerYear = 19,
  LoginExternal = 20,
  Login = 21,
  RegisterUser = 22,
  GetUserData = 23,
  GetLatestWords = 24,
  GetMostSearchedLemma = 25,
  GetMostSearchedRoot = 26,
  // ConfirmUserEmail = 27,
  ResetPasswordRequest = 28,
  ResetPassword = 29,
  ChangePassword = 30,
  GetFileByRootId = 31,
  GetReferenceCover = 32,
  GetRootAutoCom = 33,
  GetRootUpPosition = 34,
  GetRootDownPosition = 35,
  RefreshToken = 36,
  ActivateUserAccount = 37,
  ReSendActivationCode = 38,
  GetBookmarkCounts = 39,
  GetLexialBookmarks = 40,
  GetDocumentBookmarks = 41,
  AddBookmark = 42,
  RemoveBookmark = 43,
  RemoveAllBookmarkByType = 44,
  GetCorpusBookmarks = 45,
  GetSearchCriteriaBookmarks = 46,
  ContactUs = 47,
  SendComment = 48,
  UpdateUserProfile = 49,
  GetProfileData = 50,
  GetEtymologicalLexicalSheetForLemma = 51,
  ValidateRecaptcha= 52,
  SendDictionaryComment= 53,
  GetParticipantGroups= 54,
  GetParticipantsByGroup= 55,
  GetLemmasCountPerDuration = 56,
}